backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp_socks==0.11.0
aioredis==2.0.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.14.3
bidict==0.23.1
black==26.1.0
boto3==1.42.58
botocore==1.42.58
brotli==1.2.0
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
contourpy==1.3.3
crewai==1.6.1
cryptography==46.0.5
cuda-bindings==13.2.0
cuda-pathfinder==1.4.4
cuda-toolkit==13.0.2
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
elevenlabs==2.39.0
email-validator==2.3.0
emergentintegrations==0.1.0
et_xmlfile==2.0.0
fal_client==0.13.1
fastapi==0.135.2
fastuuid==0.14.0
ffmpy==1.0.0
filelock==3.25.0
flake8==7.3.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gradio==6.9.0
gradio_client==2.3.0
greenlet==3.3.2
groovy==0.1.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hiredis==3.3.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
instaloader==4.15.1
instructor==1.14.5
isort==8.0.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.1.0
jq==1.11.0
json5==0.13.0
json_repair==0.25.2
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
librt==0.8.1
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matrix-nio==0.25.2
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
motor==3.3.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.29.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pathspec==1.0.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
playwright==1.58.0
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
portalocker==2.7.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-jose==3.5.0
python-multipart==0.0.22
python-socketio==5.16.1
python-socks==2.8.1
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.26.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
runwayml==4.8.0
s3transfer==0.16.0
s5cmd==0.2.0
safehttpx==0.1.7
scipy==1.17.1
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.3
starlette==0.52.1
stripe==14.4.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.26
ultralytics-thop==2.0.18
unpaddedbase64==2.1.0
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.12
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.0
boto3==1.42.58
httpx==0.28.1
httpx-sse==0.4.3
safehttpx==0.1.7
bcrypt==4.1.3
PyJWT==2.11.0
